ai-packages/agentic-rag-module/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arize-phoenix-client==1.26.0
arize-phoenix-evals==2.6.0
arize-phoenix-otel==0.14.0
arize-phoenix==12.17.0
asttokens==3.0.0
async-property==0.2.2
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dataclasses-json==0.6.7
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
Events==0.5
executing==2.2.1
fastapi==0.115.11
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.117.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.39.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
gritql==0.2.0
grpc-google-iam-v1==0.14.2
grpcio-status==1.71.0
grpcio-tools==1.71.0
grpcio==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx-sse==0.4.1
httpx==0.28.1
ibm_watsonx_ai==1.3.39
ibm-cos-sdk-core==2.13.5
ibm-cos-sdk-s3transfer==2.13.5
ibm-cos-sdk==2.13.5
idna==3.10
ipython_pygments_lexers==1.1.1
ipython==9.6.0
jedi==0.19.2
jiter==0.11.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jwcrypto==1.5.6
langchain-cli==0.0.36
langchain-community==0.3.20
langchain-core==0.3.76
langchain-google-vertexai==2.0.15
langchain-ibm==0.3.8
langchain-ollama==0.2.3
langchain-openai==0.3.9
langchain-text-splitters==0.3.7
langchain==0.3.21
langgraph-checkpoint==2.1.1
langgraph-opensearch==0.1.5
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langgraph==0.6.8
langserve==0.3.2
langsmith==0.3.45
lomond==0.3.3
markdown-it-py==4.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.3
ollama==0.6.0
openai==1.109.1
openinference-instrumentation-langchain==0.1.55
openinference-instrumentation==0.1.42
openinference-semantic-conventions==0.1.25
opensearch-py==3.0.0
opensearch-py==3.0.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==24.2
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1_modules==0.4.2
pyasn1==0.6.1
pycparser==2.23
pydantic_core==2.33.2
pydantic-settings==2.11.0
pydantic==2.11.9
Pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests-toolbelt==1.0.0
requests==2.32.3
rich==14.1.0
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==1.8.2
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.11.0
tomlkit==0.13.3
tqdm==4.67.1
traitlets==5.14.3
typer==0.19.2
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.1.0
uvicorn==0.37.0
wcwidth==0.2.14
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
